+++ core/base-service/base.js
@@ -47,10 +47,6 @@ const optionalStringWhenNamedLogoPresent = Joi.alternatives().conditional(
   },
 )
 
-const optionalNumberWhenAnyLogoPresent = Joi.alternatives()
-  .conditional('namedLogo', { is: Joi.string().required(), then: Joi.number() })
-  .conditional('logoSvg', { is: Joi.string().required(), then: Joi.number() })
-
 const serviceDataSchema = Joi.object({
   isError: Joi.boolean(),
   label: Joi.string().allow(''),
@@ -65,7 +61,7 @@ const serviceDataSchema = Joi.object({
   namedLogo: Joi.string(),
   logoSvg: Joi.string(),
   logoColor: optionalStringWhenNamedLogoPresent,
-  logoWidth: optionalNumberWhenAnyLogoPresent,
+  logoSize: optionalStringWhenNamedLogoPresent,
   cacheSeconds: Joi.number().integer().min(0),
   style: Joi.string(),
 })
@@ -154,12 +150,18 @@ class BaseService {
   static get _cacheLength() {
     const cacheLengths = {
       build: 30,
-      license: 3600,
-      version: 300,
       debug: 60,
-      downloads: 900,
-      rating: 900,
-      social: 900,
+
+      'platform-support': 300,
+      size: 300,
+      version: 300,
+
+      chat: 1800,
+      downloads: 1800,
+      rating: 1800,
+      social: 1800,
+
+      license: 14400,
     }
     return cacheLengths[this.category]
   }