+++ services/pub/pub-popularity.tester.js
@@ -1,23 +1,12 @@
-import { isIntegerPercentage } from '../test-validators.js'
-import { createServiceTester } from '../tester.js'
+import { ServiceTester } from '../tester.js'
 
-export const t = await createServiceTester()
-
-t.create('pub popularity (valid)').get('/analysis_options.json').expectBadge({
-  label: 'popularity',
-  message: isIntegerPercentage,
+export const t = new ServiceTester({
+  id: 'PubPopularity',
+  title: 'PubPopularity',
+  pathPrefix: '/pub/popularity',
 })
 
-t.create('pub popularity (not found)')
-  .get('/analysisoptions.json')
-  .expectBadge({
-    label: 'popularity',
-    message: 'not found',
-    color: 'red',
-  })
-
-t.create('pub popularity (invalid)').get('/analysis-options.json').expectBadge({
+t.create('pub popularity').get('/analysis_options.json').expectBadge({
   label: 'popularity',
-  message: 'invalid',
-  color: 'lightgrey',
+  message: 'no longer available',
 })